requirements.txt
Django
djangorestframework==3.10.0
pytest==5.3.5
pytest-django==3.8.0
psycopg2==2.8.4
gunicorn==20.0.4
